requirement.txt
asgiref==3.8.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cryptography==42.0.5
defusedxml==0.7.1
Django==5.1
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
gTTS==2.3.1
idna==3.4
MouseInfo==0.1.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.0
pandas==1.5.3
Pillow==9.4.0
playsound==1.2.2
PyAutoGUI==0.9.53
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pywin32==305
razorpay==1.3.0
requests==2.28.2
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==2.2.0
tzdata==2022.7
urllib3==1.26.14
wikipedia==1.4.0
